🎨 : api용 interface 파일 생성
